feat(publisher): Introduce Telemetry
This change consolidates all things related to metrics and logging in a dedicated
Telemetry process to give room for processing performance when publishing.
