+++ .env.sample
@@ -11,6 +11,7 @@ SURREALDB_URL=127.0.0.1:8000
 SURREALDB_USER=root
 SURREALDB_PASS=root
 CONCURRENCY_LIMIT=16
+USE_ELASTIC_LOGGING=false
 ELASTICSEARCH_URL=http://127.0.0.1:9200
 ELASTICSEARCH_USERNAME=elastic
 ELASTICSEARCH_PASSWORD=generated-secret
