+++ crates/fuel-streams-publisher/Cargo.toml
@@ -18,7 +18,6 @@ actix-web = { workspace = true }
 anyhow = { workspace = true }
 async-nats = { workspace = true }
 async-trait = { workspace = true }
-bytes = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true }
 derive_more = { version = "1.0", features = ["full"] }
@@ -34,7 +33,6 @@ fuel-core-types = { workspace = true }
 fuel-streams = { workspace = true, features = ["test-helpers"] }
 fuel-streams-core = { workspace = true, features = ["test-helpers"] }
 futures = { workspace = true }
-num_cpus = "1.16"
 parking_lot = { version = "0.12", features = ["serde"] }
 prometheus = { version = "0.13", features = ["process"] }
 rand = { workspace = true }
