+++ crates/fuel-streams-publisher/src/lib.rs
@@ -11,13 +11,12 @@ mod packets;
 mod publisher;
 
 pub mod cli;
-pub mod elastic;
 pub mod identifiers;
-pub mod metrics;
 pub mod server;
+pub mod server_state;
 pub mod shutdown;
-pub mod state;
-pub mod system;
+
+pub mod telemetry;
 
 use std::{env, sync::LazyLock};
 
@@ -32,8 +31,6 @@ pub static CONCURRENCY_LIMIT: LazyLock<usize> = LazyLock::new(|| {
         .and_then(|val| val.parse().ok())
         .unwrap_or(32)
 });
-
-pub const FUEL_ELASTICSEARCH_PATH: &str = "fuel-data-systems";
 
 pub fn sha256(bytes: &[u8]) -> Bytes32 {
     let mut sha256 = Sha256::new();
