+++ crates/fuel-streams-publisher/src/packets.rs
@@ -1,15 +1,10 @@
 use std::sync::Arc;
 
-use elasticsearch::params::Refresh;
 use fuel_streams_core::prelude::*;
 use thiserror::Error;
 use tokio::{sync::Semaphore, task::JoinHandle};
 
-use crate::{
-    elastic::{ElasticSearch, ElasticSearchError},
-    metrics::{publish_with_metrics, PublisherMetrics},
-    FUEL_ELASTICSEARCH_PATH,
-};
+use crate::telemetry::Telemetry;
 
 #[derive(Error, Debug)]
 pub enum PublishError {
@@ -17,8 +12,6 @@ pub enum PublishError {
     StreamPublish(String),
     #[error("Semaphore acquisition failed: {0}")]
     Semaphore(#[from] tokio::sync::AcquireError),
-    #[error("Failed to log to ElasticSearch: {0}")]
-    ElasticSearchLog(#[from] ElasticSearchError),
     #[error("Unknown error: {0}")]
     Unknown(String),
 }
@@ -26,11 +19,10 @@ pub enum PublishError {
 #[derive(Clone)]
 pub struct PublishOpts {
     pub semaphore: Arc<Semaphore>,
-    pub metrics: Arc<PublisherMetrics>,
     pub chain_id: Arc<ChainId>,
     pub block_producer: Arc<Address>,
     pub block_height: Arc<BlockHeight>,
-    pub elastic_logger: Option<Arc<ElasticSearch>>,
+    pub telemetry: Arc<Telemetry>,
 }
 
 // PublishPacket Struct
@@ -63,7 +55,7 @@ impl<T: Streamable + 'static> PublishPacket<T> {
         let payload = Arc::clone(&self.payload);
         let subject = Arc::clone(&self.subject);
         let wildcard = self.wildcard;
-        let elastic_logger = opts.elastic_logger.clone();
+        let telemetry = Arc::clone(&opts.telemetry);
 
         tokio::spawn(async move {
             let _permit = opts
@@ -72,46 +64,33 @@ impl<T: Streamable + 'static> PublishPacket<T> {
                 .await
                 .map_err(PublishError::Semaphore)?;
 
-            // Publish to stream
-            publish_with_metrics(
-                stream.publish(&*subject, &payload),
-                &opts.metrics,
-                &opts.chain_id,
-                &opts.block_producer,
-                wildcard,
-            )
-            .await
-            .map_err(|e| PublishError::StreamPublish(e.to_string()))?;
+            match stream.publish(&*subject, &payload).await {
+                Ok(published_data_size) => {
+                    telemetry.log_info(&format!(
+                        "Successfully published for stream: {}",
+                        wildcard
+                    ));
+                    telemetry.update_publisher_success_metrics(
+                        wildcard,
+                        published_data_size,
+                        &opts.chain_id,
+                        &opts.block_producer,
+                    );
 
-            // Log to ElasticSearch after successful publish
-            if let Some(elastic_logger) = elastic_logger {
-                Self::log_to_elasticsearch(elastic_logger, payload, subject)
-                    .await?;
+                    Ok(())
+                }
+                Err(e) => {
+                    telemetry.log_error(&e.to_string());
+                    telemetry.update_publisher_error_metrics(
+                        wildcard,
+                        &opts.chain_id,
+                        &opts.block_producer,
+                        &e.to_string(),
+                    );
+
+                    Err(PublishError::StreamPublish(e.to_string()))
+                }
             }
-
-            Ok(())
         })
-    }
-
-    async fn log_to_elasticsearch(
-        elastic_logger: Arc<ElasticSearch>,
-        payload: Arc<T>,
-        subject: Arc<dyn IntoSubject>,
-    ) -> Result<(), PublishError> {
-        let id = subject.parse();
-        elastic_logger
-            .get_conn()
-            .index(
-                FUEL_ELASTICSEARCH_PATH,
-                Some(&id),
-                &*payload,
-                Some(Refresh::WaitFor),
-            )
-            .await
-            .map_err(|e| {
-                tracing::error!("Failed to log to ElasticSearch: {:?}", e);
-                e
-            })?;
-        Ok(())
     }
 }
